pool: fix startup for pools that saved configuration earlier than 2.2
A recent patch introduced strict checking of command options: if a
command supplies an option that is not recognised then the command will
fail.  This is limited to annotated commands.

Pools save their configuration in a 'setup' file as a series of
commands that are to be run on startup.  Unfortunately, this includes a
command with an option that isn't currently accepted. The result is
that pools with 'setup' files created earlier than 1.9.13 do not start.

This patch fixes the problem by updating the command to accept the
option but ignore it.  This is functionally the same as before
introducing the strict checking.

This patch also updates the pool defaults.  This is a set of commands
that are always executed before 'setup' to provide a base-line for
pools.  The current defaults includes the problematic command.

FOR RELEASE NOTES:

A change included in 2.6.8 prevented pools from starting up where the
pool has a 'setup' file created with a version of dCache before
v1.9.13.  This patch fixes this issue.

Target: master
Request: 2.7
Request: 2.6
Ticket: http://rt.dcache.org/Ticket/Display.html?id=8006
Requires-notes: yes
Requires-book: no