srm: Resolve regression in PostgreSQL compatibility
PostgreSQL doesn't support updatable RowSets unless the table
got a primary key. Thus RequestPropertyStorage is broken as
it relies on an updatable ResultSet for a table without a primary
key.

This patch restores the old logic relying on two statements. The
entire operation is wrapped in a TransactionTemplate to ensure
atomicity of the update.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6006/