srm: Introduce srm-common module and rename existing srm modules
Addresses bloat in the srmclient packages.

srmclient relies on srm for protocol and utility classes. The dependency
causes the srmclient packages to include lots of transitive dependencies
not actually needed by the client. Those dependencies increase the size
of the RPM and increase the startup time of the clients as the classpath
is longer.

This patch moves those classes to a new srm-common module. The srmclient
module is renamed to srm-client and the srm module to srm-server; this
follows the naming scheme used by other modules.

The patch also restricts the scope of some powermock components to
testing. This avoid that these are included in the final product.

The effect of the pathc is that the RPM shrinks from a bit above 16 MB
to sligthly below 8 MB.

The patch removes a few old build files and build instructions that
no longer apply.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6024/