srm: Fix srmAbortFiles and srmAbortRequest implementations
srmAbortFiles and srmAbortRequest are not at all compliant with the
SRM 2.2 spec. One symptom is that one may get failurs like

23 Sep 2013 23:06:57 (SRM-dcache-vm) [127.0.0.1:39844 t+Q:1:srm2:abortFiles:-2147482647] Illegal State Transition : Illegal state transition from Done to Canceled

when aborting requests that have already completed.

Another big fault is that the SURL return status used is that of
the file request, whereas the spec defines it to be the return
status of the abort operation, with only three legal return
values.

Finally, the return status of the entire request doesn't reflect the
return status of the individual SURLs.

These problems are addresses by this patch. A utility method for computing
a summary return status for the request has been added and a number of
otherwise unrelated classes are modified to make use of the utility method.

Target: trunk
Require-book: no
Require-notes: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/6040/