Upgrade BoneCP and Guava
BoneCP and Guava have to be upgraded together as the new BoneCP uses the new
Guava, and the new Guava is incompatible with the old BoneCP.

BoneCP 0.7 contains a bug in which it doesn't restore the auto-commit flag
on recycled connections. We work arround that problem in our code by carefully
resetting the flag before 'closing' the connection, however that only works
as long as you don't mix your code with eg DataNucleus. DataNucleus changes
the auto commit flag and thus you run into problems if your own code relies
on connections haven't auto commit enabled by default (as required by the
JDBC spec).

The problem is fixed in BoneCP 0.8, although that is only available in a release
candidate. rc3 has however been adopted by the Play project, which I consider
a high profile endorsement.

0.8-rc3 contains a bug in that the default auto-commit value is false rather
than true as required by the standard. I add a custom bonecp-config.xml to
work around that problem. The config also enables logging of slow SQL statements,
using a conservative 10 second threshold (this is the only thing worth
mentioning in the release notes).

The BoneCP setReleaseHelperThreads method is deprecated in BoneCP 0.8 and I
removed all calls to it.

Guava removes the Ranges class, which means a number of classes had to be
updated to compensate.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Karsten Schwank <karsten.schwank@desy.de>
Patch: http://rb.dcache.org/dashboard/