logback: revert to previous version (broken remote logging)
https://github.com/dCache/dcache/commit/94e3ae116b6f54ed90f80b07a05f7d836f8b97bc

in which the logback library version was bumped from 1.0.12 to 1.0.13 breaks alarms because remote logging hangs.

Until we ascertain why the socket appender does not work with the new version, this patch switches back to the previous version.

Update:  this is a known bug which is not yet resolved:

http://jira.qos.ch/browse/LOGBACK-896?page=com.atlassian.jira.plugin.system.issuetabpanels:changehistory-tabpanel

Testing:
Before patch, dcache alarm send command hangs.
After patch, it succeeds.

Target: master
Request: 2.7
Require-book: no
Require-notes: no
Acked-by: Gerd