(2.2) billing (plots): adjust upper bound of aggregate plots
Because of the fact that the aggregate tables are now updated only once a day for the data from the previous day, it does not make sense to round the current date to the next 24-hour-period to give the upper bound for the plots of those tables.

Testing:

Redeployed and observed plots.

Target: 2.2
Patch: http://rb.dcache.org/r/6036/
Require-notes: no
Require-book: no
Acked-by: Karsten