xrootd: Make xrootdPlugins unscoped for door
The consequences of accidentally configuring xrootdPlugins in
dcache.conf can be quite big: Access to the door would be
unauthenticated, and possible name space mapping plugins would
not be applied.

To reduce the risk of this problem, and since we removed the
scoping operator in 2.7 anyway, this patch makes xrootdPlugins
unscoped for the door. Pools continue to use the scoped xrootdPlugins
property, but the patch introduces the pool.mover.xrootd.plugins
name as an alternative. This is the same name that was introduced
in 2.7.

Target: 2.6
Require-notes: yes
Require-book: yes