srm: Clean up and fix permission operations
Addresses issues in the implementation of srmCheckPermission,
srmGetPermission, and srmSetPermission, all of which failed to
use the proper error codes.

Addresses an issue in srmSetPermission in which permission bits
where toggled rather than cleared. We now report that the REMOVE
operation is not supported as the semantics described in the SRM
spec only is applicable to ACLs.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6094/