(2.6) dcache, webadmin: fix exception handling
This commit covers two patches:

    http://rb.dcache.org/r/6032/
    http://rb.dcache.org/r/6033/

BILLING:

A number of checked exceptions which were wrapping unchecked exceptions have been eliminated.  In general RuntimeExceptions are propagated. Two exceptions:  when generating JAIDA plots, and wit

CORE, XACML, WEBADMIN:

Code in several places is not handling exceptions correctly; in particular, RuntimeExceptions masked by Exception are occasionally being caught and not propagated.

Where it is possible, catch(Exception) has been eliminated; where not possible, catch(RuntimeException) is called first so that this exception can be rethrown.

In several cases, a RuntimeException (such as the JDOExceptions from DataNucleus) are not propagated, as they cannot reasonably be handled elsewhere, and they should not be fatal to the JVM in

A separate set of patches have been reviewed for 2.6.

Testing:

Junit tests rerun.  Deployed and checked.

Target: 2.6
Acked-by: Karsten
Require-notes: no
Require-book: no