dcache-webadmin: fix table sorting for all relevant pages
This patch addresses both a general and specific problem.  Sorting of tables via javascript was not working universally.  This depended on the following sub-issues:

1.  The change to Wicket 1.6 also changed the order in which script is added to the header (as well as the underlying Java APIs for dealing with it); in its current state, this was producing inconsistent nesting of scripts which the browser could not handle.

2.  The sorting code in javascript depended on css id (#), but if there is more than one table per page, this does not work for the subsequent tables.

3.  Not all relevant tables were tagged correctly.

To address the first issue, some refactoring was done.  BasePage now has a subclass, SortableBasePage.  The base page provides an adapter inner method for rendering the page head markup which should be overridden (always calling super first) with the specific script or references for that subclass.  The page itself should provide all the necessary head material for any of the panels it may use.

The second issue is fixed by altering the css and the javascript to use class instead of id.

The final issue was fixed by adding the necessary "class=sortable" marker to the table.

A separate patch will be necessary for 2.6.

The bug reference below is really more of a feature request which is only partially handled here.

Testing:

Clicking around by hand.

Target: master
Patch: http://rb.dcache.org/r/6083
Bug: http://rt.dcache.org/Ticket/Display.html?id=7938
Require-notes: yes
Require-book: no
Acked-by: Gerd

RELEASE NOTES:
Fixes issues preventing the sorting of tables and subtables in webadmin.