srm: Fix regression in srmGetRequestSummary
Addresses a regression in srmGetRequestSummary that caused
the return status of the summary request to be computed from
the status of the requests that were queried.

The spec is not very precise about what constitutes a failure
or partial failure of the query, but I chose to consider not
finding a matching request token a failure of the query itself.

I also updated the client to report file level request information
even if the query return status is a failure.

Target: trunk
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/6153/