srm: Use description of last job transition in return status messages
Improves status messages returned to SRM clients.

The SRM in some places uses the internal state as a description text,
and in other places the current "error message", which is the description
of the last state transition with an added time stamp. In ls it even
uses the concatenated description messages of all state transitions.

Since this information should probably not leak to the outside (its
confusing and it makes FTS logs harder to understand), this patch uses
the description text of the last transition itself.

In subsequent patches I will try to improve the description messages
in calls to setState().

A consequence of this patch is that it is no longer necessary to maintain
the concatenated errorMessage field in the Job class - it is no longer
used.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Patch: http://rb.dcache.org/r/6156/