chimera: Provide meaningful exceptions on failure to delete a file
Addresses issues in Chimera in which failures to delete a file or
directory result in a generic error message without information
about the actual error.

The root cause is that Chimera copies a bad API from Java in which
remove methods simply return a boolean indicating whether the file
was deleted or not. This patch changes this such that an exception
is thrown in case of failure.

Motivation for this patch can be found in server logs of running
S2 against dCache (see http://svn.dcache.org/build/job/master%20post%20testing/167/console).
The log contains entries like:

30 Oct 2013 23:07:14 (PnfsManager) [RiY:547:srm2:rmdir SRM-sisyphus PnfsDeleteEntry] Failed to delete entry: Entry could not be removed: /data/s2/vm-dcache-wn01-tester5/20131030-230600-25706x

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Acked-by: Tigran Mkrtchyan <tigran.mkrtchyan@desy.de>
Patch: http://rb.dcache.org/r/6169/