pnfsmanager: Fix error type when delete is refused
Addresses an issue that causes

30 Oct 2013 23:07:32 (PnfsManager) [RiY:610:srm2:rm SRM-sisyphus PnfsDeleteEntry] Inconsistent message class diskCacheV111.vehicles.PnfsDeleteEntryMessage received form [>SRM-sisyphus@dCacheDomain]

to be logged in PnfsManager.

PnfsManager uses INVALID_ARGS as an internal error. Hence using
it to signal that the allowed file types do not match the actual
file type leads to the above error. The patch changes the logic
to other error codes. The error codes may not be optimal in all
cases (eg for links or special files), but they are better than
the existing error code.

Target: trunk
Require-notes: yes
Require-book: no
Acked-by: Dmitry Litvintsev <litvinse@fnal.gov>
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6171/