dcache-webadmin: fix regression (submit failure) on ActiveTransfers introduced by auto-refresh
After application of http://rb.dcache.org/r/6085 it was discovered that on pages where administrative interaction occurs, in particular PoolList and ActiveTransfers, the addition of the Ajax autorefresh behavior interfered with the ability to issue commands to the pool.

The reason was that auto-refresh was swapping out the object on the page but the new object was not being propagated to the underlying list view.  Hence checking the checkbox in order to select the row was registered on the new object, but not the old, which was still being held onto in the table model.

Two fixes were necessary.

1.  Turn off the prior call to the pool DAO when form submit is taking place, allowing the old model object to be used.  The flag is reset so that refresh will occur at the next timeout.
2.  Move the refresh call to inside the panel's list view "onBeforeRender" method.  The Ajax beforeRender(), which calls the refresh() adapter for the page, is not useful here, as it follows, rather than precedes, the beforeRender of the panel component.

Since information does not propagate back immediately from the pool manager, a "pending" marker has been added to allow the user to see that the command has been issued but that the state change is pending.  If the user does not manually refresh the page (usually a few seconds wait is sufficient), the new state from the pool manager will appear at the next auto-refresh (after 1 minute).  In the meantime, the entry has its state marked as "state + (pending)" and the row is colored red.

This patch only applies to ActiveTransfers.  A separate patch for PoolList (6161) has been posted.

Testing Done:

Before application of the patch, no admin commands can be issued from the "ActiveTransfer" page (kill movers).

After the patch, commands are issued and the page is correctly updated.

Tested by:

Starting several long-lived concurrent writes to pools.
Waiting for mover info to appear.
Selecting a mover and canceling it.
Waiting to see final update/disappearance of mover from list on refresh

Target: master
Patch: http://rb.dcache.org/r/6162
Requires-notes: no
Requires-book: no
Acked-by: Tigran