gplazma: Avoid multiple primary FQANs
Addresses a bug that can cause multiple primary FQANs to be created,
thus resulting in failures in billing and spacemanager.

The vorolemap plugin may cause multiple primary FQANs to be created.
This results in SrmSpaceManager failures, eg:

06 Nov 2013 08:51:38 (SRM-gonzo) [130.235.149.204:59894 ZR0:11817:srm2:prepareToPut:-2060713576:-2060713574] Failed to mark space as being used space: Marking Space as Being Used failed
=>Subject has multiple primary FQANs

This is a regression introduced in dCache 2.6 when we altered the
gPlazma plugin interface. Before, only the FQAN for which we
generated a mapping was included as a primary FQAN.

In the fix I decided to only include the actual primary FQAN and
not add "fake" primary FQAN based on mappings that exist. This is
a change compared to pre-2.6 semantics, but I believe this is the
correct thing to do and is also more stable when multiple mapping
plugins are used.

Target: trunk
Request: 2.6
Require-book: no
Require-notes: yes
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6198/
(cherry picked from commit bc88df103b80b342be23515b8cfeeac0846ca2ce)