info-provider: Fix publishing of SRM
When srm-loginbroker was removed as a service, publishing the
srm to loginbroker was altered to publish the family name as
'srm' rather than 'SRM'. This is in line with our other doors,
but I was unaware that info-provider checks for this particular
string and that the check is case sensitive.

This patch updates info-provider to check for the lower case
string.

Target: trunk
Request: 2.7
Require-notes: no
Require-book: no
Acked-by: Paul Millar <paul.millar@desy.de>
Patch: http://rb.dcache.org/r/6222/